block douban short video
douban app will automatically play all short videos on the page. I hate auto play. I want to be able to control which to play and when to play, but douban app won't give me the power to control. So I have to block the entire domain hosting the short videos.